Implemented the forward algorithm, and supervised training for HMMs
But no smoothing.  So, on the twitter POS tagging data, this doesn't really work at all, because
almost all of the tweets have unseen words, which get assigned zero probability.

But the point of this was just to have implemented some basic algorithms, so I don't really care
to implement smoothing for this.  After implementing forward-backward for unsupervised training,
I'll move on to CRFs and neural sequence models.